requirements.txt
setuptools
PyYaml
colorlog
psutil
pybase64
boto3
azure-identity
azure-keyvault-secrets
pykeystore
pytest-mock
pytest-watch
pytest